Made some changes to OpinionMap and MoveCalculator.

+++ src/main/scala/org/kapunga/tgoban/engine/OpinionMap.scala
@@ -3,6 +3,5 @@ package org.kapunga.tgoban.engine
 import org.kapunga.tgoban.MetaBoard
 
 
-class OpinionMap(size: Int) extends MetaBoard[Float](size, 0.0) {
-
+class OpinionMap(size: Int) extends MetaBoard[Double](size, 0.0) {
 }
